Removed Shippable mixin, renamed legacy models
