+++ src/apps/products/models/course.py
@@ -1,4 +1,6 @@
 from datetime import timedelta
+from decimal import Decimal
+from typing import TYPE_CHECKING
 
 from django.apps import apps
 from django.core.exceptions import ValidationError
@@ -7,10 +9,14 @@ from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
 from apps.mailing.tasks import send_mail
-from apps.products.models.base import Shippable
+from apps.studying import shipment_factory as ShipmentFactory
 from apps.users.models import User
 from core.files import RandomFileName
-from core.models import models
+from core.models import TimestampedModel, models
+from core.pricing import format_old_price, format_price
+
+if TYPE_CHECKING:
+    from apps.orders.models import Order
 
 
 class CourseQuerySet(QuerySet):
@@ -50,11 +56,30 @@ class CourseQuerySet(QuerySet):
 CourseManager = models.Manager.from_queryset(CourseQuerySet)
 
 
-class Course(Shippable):
+class Course(TimestampedModel):
     objects = CourseManager()
 
+    name = models.CharField(max_length=255)
     name_genitive = models.CharField(_("Genitive name"), max_length=255, help_text="«мастер-класса о TDD». К примеру для записей.")
+    name_receipt = models.CharField(
+        _("Name for receipts"), max_length=255, help_text="«посещение мастер-класса по TDD» или «Доступ к записи курсов кройки и шитья»"
+    )
+    full_name = models.CharField(
+        _("Full name for letters"),
+        max_length=255,
+        help_text="Билет на мастер-класс о TDD или «запись курсов кройки и шитья»",
+    )
+    name_international = models.CharField(_("Name used for international purchases"), max_length=255, blank=True, default="")
 
+    group = models.ForeignKey("products.Group", verbose_name=_("Analytical group"), on_delete=models.PROTECT)
+    slug = models.SlugField(unique=True)
+
+    price = models.DecimalField(max_digits=8, decimal_places=2)
+    old_price = models.DecimalField(max_digits=8, decimal_places=2, blank=True, null=True)
+
+    tinkoff_credit_promo_code = models.CharField(
+        _("Fixed promo code for tinkoff credit"), max_length=64, blank=True, help_text=_("Used in tinkoff credit only")
+    )
     welcome_letter_template_id = models.CharField(
         _("Welcome letter template id"), max_length=255, blank=True, null=True, help_text=_("Will be sent upon purchase if set")
     )
@@ -90,6 +115,38 @@ class Course(Shippable):
         verbose_name = _("Course")
         verbose_name_plural = _("Courses")
         db_table = "courses_course"
+
+    # Methods from Shippable mixin
+    def get_price_display(self) -> str:
+        return format_price(self.price)
+
+    def get_old_price_display(self) -> str:
+        return format_price(self.old_price)
+
+    def get_formatted_price_display(self) -> str:
+        return format_old_price(self.old_price, self.price)
+
+    def ship(self, to: User, order: "Order") -> None:
+        return ShipmentFactory.ship(self, to=to, order=order)
+
+    def unship(self, order: "Order") -> None:
+        return ShipmentFactory.unship(order=order)
+
+    def get_price(self, promocode: str | None = None) -> Decimal:
+        promocode_obj = apps.get_model("orders.PromoCode").objects.get_or_nothing(name=promocode)
+
+        if promocode_obj is not None:
+            return promocode_obj.apply(self)
+
+        return self.price
+
+    def get_template_id(self) -> str | None:
+        """Get custom per-item template_id"""
+        if not hasattr(self, "template_id"):
+            return None
+
+        if self.template_id is not None and len(self.template_id):
+            return self.template_id
 
     def clean(self) -> None:
         """Check for correct setting of confirmation_template_id and confirmation_success_url"""
