+++ src/apps/products/tests/products/bundles/tests_bundle_deprecated.py
@@ -5,4 +5,4 @@ pytestmark = [pytest.mark.django_db]
 
 def test(mixer):
     with pytest.raises(RuntimeError, match="Deprecated"):
-        mixer.blend("products.Bundle")
+        mixer.blend("products.LegacyBundle")
