+++ src/apps/products/tests/products/records/tests_record_deprecated.py
@@ -5,4 +5,4 @@ pytestmark = [pytest.mark.django_db]
 
 def test(mixer):
     with pytest.raises(RuntimeError, match="Deprecated"):
-        mixer.blend("products.Record")
+        mixer.blend("products.LegacyRecord")
